+++ .gitignore
@@ -328,6 +328,3 @@ ASALocalRun/
 
 # MFractors (Xamarin productivity tool) working folder 
 .mfractor/
-/OECandS_Lab№1
-/AandDSlab1
-/AandDS/AandDS
