--- Lab1/Lab1/IAircraft.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Lab1
-{
-    public interface IAircraft
-    {
-        void SetPosition(int x, int y, int width, int height);
-        void MoveTransport(Direction direction);
-        void Draw(Graphics g);
-    }
-}
