+++ Lab1/Lab1/Lab1.csproj
@@ -53,11 +53,8 @@
     <Compile Include="FormFighter.Designer.cs">
       <DependentUpon>FormFighter.cs</DependentUpon>
     </Compile>
-    <Compile Include="IAircraft.cs" />
-    <Compile Include="Plane.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="War_plane.cs" />
     <EmbeddedResource Include="FormFighter.resx">
       <DependentUpon>FormFighter.cs</DependentUpon>
     </EmbeddedResource>
